moved multi-threaded executor functions in seperate file
